API/requirements.txt
Flask==2.2.3
flask-cors==3.0.10
flask-restful==0.3.9
flask-sqlalchemy==3.0.3
tensorflow==2.11.0
keras==2.11.0
librosa==0.10.0
scikit-learn==1.2.2
joblib==1.2.0
python-dotenv==1.0.0
Flask-JWT-Extended==4.4.4
python-decouple==3.8
bcrypt==3.2.0
psycopg2-binary==2.9.6
